python-ai/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
flask>=3.0.0
flask-cors>=4.0.0
ortools>=9.8.0
requests>=2.31.0
python-dotenv>=1.0.0
